requirements.txt
Django==2.0.3
django-macaddress==1.5.0
django-versionfield2==0.5.0
netaddr==0.7.19
pytz==2018.3
six==1.11.0
